# WebApplication1/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApplication1",
    "version": "v1"
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": [
          "products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/product"
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/product"
            }
          }
        }
      },
      "put": {
        "tags": [
          "products"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "product": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "brand": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "expiredDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}